Reworked /stats/**
